chore: Rename analytics -> googleAnalytics
